fixes with pending
