+++ app/javascript/hooks/queries/admin/manage_users/usePendingUsers.jsx
@@ -1,13 +1,8 @@
 import { useQuery } from 'react-query';
-import { useSearchParams } from 'react-router-dom';
 import axios from '../../../../helpers/Axios';
 
 export default function usePendingUsers(input, page) {
-  const [searchParams] = useSearchParams();
-
   const params = {
-    'sort[column]': searchParams.get('sort[column]'),
-    'sort[direction]': searchParams.get('sort[direction]'),
     search: input,
     page,
   };
